Change default min interval to 150ms
While on Linux 100ms seems to be working concistenlty. Most people will
be on Windows so changing the hints to that. This will make stripts for
Linux users a little bit more boring, but they will consistnly work
across platforms. The needs of the many outway the needs of the few ;-)