Add an funscripting blender setup file
My blender funscripting setup uses with two scenes. Loading a movie in
the importmovie scene. Build proxies and switch to the scripting scene.
There the imported scene is references in 3 preview windows with each a
frame offset. Make sure the FPS of both scenes match.

Script in the bottom strip which has the actual timing of the movie.